Avoid conversion to float when 'unit' is supplied but empty (fix for #22).
Some webbox supplies an empty 'unit' property for non-numerical sensors.
Up to now, code was using availability of the 'unit' property in data to
try a float conversion that lead to None/an 'unknown' HA state when data
was not numerical.

Modified code is using the _attr_unit_of_measurement property that is
defined when 'unit' matches one of the predefined webbox unit.