Handle numeric sensors without unit_of_measurement attribute when they report a non-numeric value
When offline, SMA webbox reports a string (usually '---') for measurement sensors.

This was correctly handled for sensor with a unit_of_measurement attribute but not for others, e.g. the resistance unit doesn't exist in HomeAssistant.

For such a sensor, the original string was not replaced with "None" (interpreted as "unknown") and returned to HA, resulting in an exception in the HA code that assumed a number...

The fix is now applied to all sensor with a state_class attribute (measurements)